Optimize inter frame gen to minimize use of dupes
This commit refactors, adds logs debugging information, and fixes
multiple problems with the render module - most importantly an
issue where Butterflow was not making enough interpolated frames.
This led to duping the final frame in every segment to compensate
for missing frames.

Another major bug that was discovered and fixed was that sometimes
the time step was set to a value that caused it to make a copy of
the next frame. This would cause the appearance of faux-dupes that
appeared right after another.