Ditch mklist when setting opencl link args in os x
The OpenCL framework isn't getting picked up because mklist
doesn't retain ordering of items that are sent to it because it
first adds the items in a set and then returns a list (sets are
unordered of course). `-arch` arg was coming in between
`-framework` and `OpenCL` in the resulting clang call.