Drop any and all future support for avconv
Ubuntu 15.04 (Vivid Vervet) provides support for ffmpeg and
Debian (Stretch) will switch back to it from avconv. Supporting two
kinds of libav providers is too much of a hassle and besides,
ffmpeg is better than avconv in almost every imaginable way which
is a conclusion reached by the Debian maintainers.