Prefer to drop interpolated frames over source
This new code will try to drop interpolated over real frames when
they are available. In addition to enhancing the image of the
output video, it optimizes the time step of interpolated frames
which evens the jumpiness between frames and is more efficent
in terms of not generating too many extra frames.