Software-only rendering
This commit introduces a pathway for people to use the app without
needing to set up OpenCL. A new module that does software frame
interpolation makes this possible.

OpenCl accelerated rendering will only be availble to those who:
set up OpenCL correctly and have a compatible device. The fastest
rendering methods will be picked automatically, however software
rendering can be forced with the new `-sw` flag.

Note that the new frame interpolation method is naievly implemented,
unoptimized and may be buggy. Results produced so far do not match
up to the OpenCL version. This will be fixed in a future commit.